do pre-check in `validate` fn
